Rewrite main loop to allow for handling multiple connections at a time
Previously, we would accept a client socket, then read the request, then
write the response, and then start again from the first step. This meant
that there could be only one socket connection at a time, which
incidentally meant that if multiple requests were sent at the same time
to the server, only one would be handled right away and the others
would be waiting in the socket's accept backlog, increasing the
response delay.

Now, we use the epoll* functions to create an event loop that can
handle a maximum of 16 socket connections at a time, still with a
single thread.

The request time out safety measure was removed to limit the size of
this change but I plan to add it back in a later commit.