Optimize memory usage and code flow in parser
The request URI and host are now packed into the same character array
(they are delimited by a NULL character). The parser state was changed
from an int field to a field that only takes 4 bits. The parser_tmp
field was packed inside of the parser state field. The parser was
edited to make all of this work.

I also edited the parser to remove some checks that are done twice and
I believe that it made following the code easier, but it might also
have made writing the code harder because more it is now a bit more
prone to mistakes.

In general, this commit allows for more characters in the URL while
keeping the same memory usage at the cost of code that might be harder
to reason about.

Finally, I've put the functions in server.c below the main function,
and changed to use a buffer from a global variable every time that we
need one.