Refactor client.c
Firstly, I renamed client to conn(ection) because it really is about
a connection, not a client. This is not about the client itself but
is about the HTTP connection's state, with the request parser state.

Secondly, I hid its implementation details inside of the implemetation
file (conn.c) and the header file now presents a simple API.

Finally, I modified the epoll.c file to use the new interface.

Please, note that during the refactor, I changed my mind about the code
style: I now want comments to be written with the /* style instead of
the // style. I am therefore starting to modify the comments.