Accept any method instead of just GET
Now, we can wait indefinitely while waiting for the HTTP method if
the client never sends a space. This can be a DoS vulnerability.
However, we already had such a vulnerability because the client
could already send a infinite request line or infinite header by
never sending the CRLF characters, so this commit doesn't create a
new vulnerability IMO.